refactor articleController
