+++ controllers/userController.js
@@ -2,9 +2,8 @@ const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const config = require('config');
 
-const { check, validationResult } = require('express-validator');
+const { validationResult } = require('express-validator');
 
-// const auth = require('../middleware/auth');
 const User = require('../models/user');
 
 // @route     GET /users/me
