--- routes/auth.js
@@ -1,73 +1,0 @@
-const express = require('express');
-const bcrypt = require('bcryptjs');
-const jwt = require('jsonwebtoken');
-const config = require('config');
-
-const router = express.Router();
-const { check, validationResult } = require('express-validator');
-const auth = require('../middleware/auth');
-
-const User = require('../models/user');
-
-// @route     GET /auth
-// @desc      Get logged in user (email, name)
-// @access    Private
-router.get('/', auth, async (req, res) => {
-  try {
-    // req.user from auth middleware
-    const user = await User.findById(req.user.id).select('-password');
-    res.json(user);
-  } catch (err) {
-    console.error(err.message);
-    res.status(500).send('Server Error');
-  }
-});
-
-// @route     POST /auth
-// @desc      Auth user & get token
-// @access    Public
-router.post('/', [
-  check('email', 'Please include a valid email').isEmail(),
-  check('password', 'Password is required').exists(),
-],
-async (req, res) => {
-  const errors = validationResult(req);
-  if (!errors.isEmpty()) {
-    // change error
-    return res.status(400).json({ errors: errors.array() });
-  }
-
-  const { email, password } = req.body;
-
-  try {
-    const user = await User.findOne({ email });
-
-    if (!user) {
-      return res.status(400).json({ msg: 'Invalid Credentials' });
-    }
-    const isMatch = await bcrypt.compare(password, user.password);
-
-    if (!isMatch) {
-      return res.status(400).json({ msg: 'Invalid Credentials' });
-    }
-
-    const payload = {
-      user: {
-        id: user.id,
-      },
-    };
-
-    jwt.sign(payload, config.get('JWT_SECRET'), {
-      expiresIn: '7d',
-      // change error
-    }, (err, token) => {
-      if (err) throw (err);
-      res.json({ token });
-    });
-  } catch (err) {
-    console.error(err.message);
-    res.status(500).send('Server Error');
-  }
-});
-
-module.exports = router;
